refactor: update AccessTokenScopeSelect to use selectedScopes prop and optimize scope handling
